Update Rot_Test_PID_Bt.ino
@rangerdawavs @nicole-0 
This new branch utilizes the Gyroscope to find current angle. It has a starting angle of 0, which is why it has an offset to adjust that number to the starting angle. Communications with the raspberry have 3 modes. This allows for correction of the coefficients during testing. Instead of sending the desired angle and magnitude, the current and desired coordinates are sent and the desired angle and magnitude are calculated in the Arduino. There might be some issues related to the type of variables. I was able to compile it in the Arduino IDE but the Raspberry Pi code must be updated to match this before it can be tested in the actual prototype. Comments need to be updated.