Fixed Arduino
Added integral coefficient. Added lights for bugtesting. removed bug where angle was calculated incorrectly. added a maximum for control_Pos to prevent overheating. The integral is only calculated within a certain range of the desired angle to remove steady state error. The errorDot_Rot is averaged with the previous value to reduce the amplitude of the noise. The same is done with the control signal when the integral is calculated.